In the class addCourseCommand:
fixed a bug that duplicate course (here, duplicate means course with the same name) would be allowed to add in the system.

Also, changed the idea to check the duplicate course
(Only check the name, by having a list to store all the names of the courses that are in the system, then check if the name of the course that is trying to be added is in the list).

Now, duplicate course would not be allowed to added in the system, an error message would be given if trying to add a course with the name that is already in the system.